Refactor graphql generated code and fix naming
Moved generated code into a __generated__ folder and made naming more explicit/consistent so its clear when an Admin client is being used vs a user client.
