+++ api-lib/Gql.ts
@@ -1,7 +1,10 @@
 import './node-fetch-shim';
-import { Gql } from '../src/lib/gql/GqlHasuraAdmin';
+import { AdminClient } from '../src/lib/gql/adminClient';
 
 import { HASURA_GRAPHQL_ADMIN_SECRET, NODE_HASURA_URL } from './config';
 
-export * from '../src/lib/gql/GqlHasuraAdmin';
-export const gql = new Gql(NODE_HASURA_URL, HASURA_GRAPHQL_ADMIN_SECRET);
+export * from '../src/lib/gql/adminClient';
+export const gql = new AdminClient(
+  NODE_HASURA_URL,
+  HASURA_GRAPHQL_ADMIN_SECRET
+);
