+++ scripts/generate-zeus.sh
@@ -5,14 +5,14 @@ set -o allexport
 source .env
 set +o allexport
 
-GEN_PATH=./src/lib/gql
+GEN_PATH=./src/lib/gql/__generated__
 
 function generate() {
   # use the first argument as the path to move files to
   local MV_PATH=$GEN_PATH/$1; shift
 
   # pass the rest of the arguments to zeus
-  (set -x; zeus "$HASURA_GRAPHQL_ENDPOINT"/v1/graphql "$GEN_PATH" --ts --rq "$@")
+  (set -x; zeus "$HASURA_GRAPHQL_ENDPOINT"/v1/graphql "$GEN_PATH" --ts "$@")
   test -d "$GEN_PATH"/zeus
   test -d "$MV_PATH" && rm -r "$MV_PATH"
   mv "$GEN_PATH"/zeus "$MV_PATH"
@@ -42,8 +42,8 @@ function generate() {
   fi
 }
 
-generate zeusHasuraAdmin -h x-hasura-admin-secret:$HASURA_GRAPHQL_ADMIN_SECRET
-generate zeusUser -h x-hasura-role:user -h "authorization:generate"
+generate zeusAdmin -h x-hasura-admin-secret:$HASURA_GRAPHQL_ADMIN_SECRET
+generate zeusUser --rq -h x-hasura-role:user -h "authorization:generate"
 
 # fix formatting of generated files
 node_modules/.bin/prettier --write $GEN_PATH
